优化慢日志记录逻辑

+++ src/com.rs
@@ -256,7 +256,7 @@ pub struct ClusterConfig {
     // dead option: always 1
     pub node_connections: Option<usize>,
 
-    pub slowlog_slow_than: Option<u128>,
+    pub slowlog_slow_than: Option<u64>,
 }
 
 impl ClusterConfig {
@@ -271,7 +271,7 @@ impl ClusterConfig {
         self.fetch_interval.unwrap_or(DEFAULT_FETCH_INTERVAL_MS)
     }
 
-    pub fn set_slowlog_slow_than(&mut self, value: u128) {
+    pub fn set_slowlog_slow_than(&mut self, value: u64) {
         if self.slowlog_slow_than.is_none() {
             self.slowlog_slow_than.replace(value);
         }
